requirements.txt
fastapi
uvicorn
python-dotenv
twelvedata
finnhub-python
groq
requests
pandas